--- a/tests/diff_left.conf
+++ b/tests/diff_left.conf
@@ -409,6 +409,18 @@ type na_union_perm_source, na_union_perm_a, na_union_perm_c;
 type na_union_perm_target, na_union_perm_b;
 neverallow na_union_perm_source na_union_perm_target:infoflow { hi_w med_w low_w };
 
+attribute redundant_rule_attr;
+type redundant_rule_source_1, redundant_rule_attr;
+type redundant_rule_source_2, redundant_rule_attr;
+type redundant_rule_target_1;
+type redundant_rule_target_2;
+bool redundant_rule_bool true;
+allow redundant_rule_attr redundant_rule_target_1:infoflow low_w;
+allow redundant_rule_attr redundant_rule_target_2:infoflow med_w;
+if (redundant_rule_bool) {
+allow redundant_rule_source_2 redundant_rule_target_2:infoflow hi_w;
+}
+
 # type_transition rule differences
 type tt_matched_source;
 type tt_matched_target;