--- a/tests/diff_left_standard.conf
+++ b/tests/diff_left_standard.conf
@@ -697,6 +697,18 @@ type dax_union_perm_source, dax_union_perm_a, dax_union_perm_c;
 type dax_union_perm_target, dax_union_perm_b;
 dontauditxperm dax_union_perm_source dax_union_perm_target:infoflow ioctl { 0x1-0x3 };
 
+attribute redundant_rule_attr;
+type redundant_rule_source_1, redundant_rule_attr;
+type redundant_rule_source_2, redundant_rule_attr;
+type redundant_rule_target_1;
+type redundant_rule_target_2;
+bool redundant_rule_bool true;
+allow redundant_rule_attr redundant_rule_target_1:infoflow low_w;
+allow redundant_rule_attr redundant_rule_target_2:infoflow med_w;
+if (redundant_rule_bool) {
+   allow redundant_rule_source_2 redundant_rule_target_2:infoflow hi_w;
+}
+
 ################################################################################
 # matching typebounds
 type match_parent;