… tests/diff_left_redundant.conf