@jwcart2 2019-10-02
Fixed (I hope) the formatting issues. No other changes.
Like I mentioned before, this just changes how AV rules are processed. No changes to AVX and TE rules. The added, removed, and modified rules are converted back into an AVRule class at the end.