@jwcart2 2019-10-15
I'll work on these changes. Would you prefer for me to make a new clean pull request or make the changes on the branch for this pull request?