@pebenito 2019-10-23
I can either squash and merge, or if you like, you could force push on here to get back to a single commit with revised commit message. I'm ok either way.